feat: move sdk gen clis from core to shematics
