+++ packages/@ama-sdk/core/.eslintrc.js
@@ -9,7 +9,6 @@ module.exports = {
       'tsconfig.build.json',
       'tsconfig.builders.json',
       'tsconfig.spec.json',
-      'tsconfig.cli.json',
       'tsconfig.eslint.json'
     ],
     'sourceType': 'module'
