+++ packages/@ama-sdk/core/package.json
@@ -67,14 +67,11 @@
     "build:esm2020": "tsc -b tsconfig.build.json",
     "postbuild": "yarn cpy './package.json' dist && patch-package-json-main",
     "prepare:build:builders": "yarn cpy 'schematics/**/*.json' dist/schematics && yarn cpy 'collection.json' dist",
-    "build:cli": "tsc -b tsconfig.cli.json --pretty && yarn generate-cjs-manifest",
     "build:builders": "tsc -b tsconfig.builders.json --pretty && yarn generate-cjs-manifest",
     "prepare:publish": "prepare-publish ./dist"
   },
   "dependencies": {
     "@swc/helpers": "~0.5.0",
-    "chokidar": "^3.5.2",
-    "minimist": "^1.2.6",
     "tslib": "^2.6.2",
     "uuid": "^9.0.0"
   },
@@ -84,7 +81,6 @@
     "@angular/common": "~17.3.0",
     "@o3r/schematics": "workspace:^",
     "@schematics/angular": "~17.3.0",
-    "globby": "^11.1.0",
     "isomorphic-fetch": "^3.0.0",
     "rxjs": "^7.8.1",
     "typescript": "~5.4.2"
@@ -131,7 +127,6 @@
     "@swc/cli": "~0.3.0",
     "@swc/core": "~1.5.24",
     "@types/jest": "~29.5.2",
-    "@types/minimist": "^1.2.2",
     "@types/node": "^20.0.0",
     "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
@@ -144,7 +139,6 @@
     "eslint-plugin-jsdoc": "~48.2.1",
     "eslint-plugin-prefer-arrow": "~1.2.3",
     "eslint-plugin-unicorn": "^51.0.1",
-    "globby": "^11.1.0",
     "isomorphic-fetch": "~3.0.0",
     "jest": "~29.7.0",
     "jest-junit": "~16.0.0",
@@ -163,10 +157,5 @@
   "engines": {
     "node": ">=18.0.0"
   },
-  "schematics": "./collection.json",
-  "bin": {
-    "amasdk-clear-index": "./dist/cli/clear-index.cjs",
-    "amasdk-files-pack": "./dist/cli/files-pack.cjs",
-    "amasdk-update-spec-from-npm": "./dist/cli/update-spec-from-npm.cjs"
-  }
+  "schematics": "./collection.json"
 }
