+++ packages/@ama-sdk/core/src/fwk/index.ts
@@ -7,5 +7,4 @@ export * from './errors';
 export * from './ignore-enum.type';
 export * from './logger';
 export * from './mocks/index';
-export * from './open-api-tools-configuration';
 export * from './Reviver';
