+++ packages/@ama-sdk/core/src/utils/index.ts
@@ -4,4 +4,3 @@ export * from './ie11';
 export * from './json-token';
 export * from './mime-types';
 export * from './generic-api';
-export * from './generators';
