+++ packages/@ama-sdk/schematics/cli/clear-index.cts
