+++ packages/@ama-sdk/schematics/package.json
@@ -6,18 +6,27 @@
   },
   "description": "Swagger specification SDK Generator by schematics",
   "typings": "./src/public_api.d.ts",
+  "main": "./src/public_api.js",
+  "types": "./src/public_api.d.ts",
   "scripts": {
     "nx": "nx",
     "ng": "yarn nx",
     "test": "yarn nx test ama-sdk-schematics",
     "prepare:build:builders": "yarn cpy 'schematics/**/*.json' 'schematics/**/templates/**' 'schematics/**/*.jar' dist/schematics && yarn cpy '{package,collection,migration}.json' dist",
     "build": "yarn nx build ama-sdk-schematics",
+    "postbuild": "patch-package-json-main",
     "build:builders": "tsc -b tsconfig.builders.json --pretty && yarn generate-cjs-manifest",
     "install-swagger-cli": "mvn install:install-file -DgroupId=io.swagger -DartifactId=swagger-codegen-cli -Dversion=2.4.0-AMADEUS -Dpackaging=jar -Dfile=./schematics/resources/swagger-codegen-cli.jar",
     "build:swagger": "yarn install-swagger-cli && run-p 'build:swagger-*' 'build:openapi-*'",
     "build:openapi-typescript-gen": "mvn clean package -f ./schematics/typescript/core/openapi-codegen-typescript/pom.xml",
     "build:swagger-java-client-gen": "mvn clean package -f ./schematics/java/client-core/swagger-codegen-java-client/pom.xml",
+    "build:cli": "tsc -b tsconfig.cli.json --pretty && yarn generate-cjs-manifest",
     "prepare:publish": "prepare-publish ./dist"
+  },
+  "bin": {
+    "amasdk-clear-index": "./dist/cli/clear-index.cjs",
+    "amasdk-files-pack": "./dist/cli/files-pack.cjs",
+    "amasdk-update-spec-from-npm": "./dist/cli/update-spec-from-npm.cjs"
   },
   "ng-update": {
     "migrations": "./migration.json",
@@ -60,8 +69,11 @@
   "dependencies": {
     "@angular-devkit/core": "~17.3.0",
     "@angular-devkit/schematics": "~17.3.0",
+    "chokidar": "^3.5.2",
+    "globby": "^11.1.0",
     "js-yaml": "^4.1.0",
     "minimatch": "~9.0.3",
+    "minimist": "^1.2.6",
     "rxjs": "^7.8.1",
     "semver": "^7.5.2",
     "tslib": "^2.6.2"
@@ -83,6 +95,7 @@
     "@types/jest": "~29.5.2",
     "@types/js-yaml": "^4.0.5",
     "@types/minimatch": "^5.1.2",
+    "@types/minimist": "^1.2.2",
     "@types/node": "^20.0.0",
     "@types/pid-from-port": "^1.1.0",
     "@types/semver": "^7.3.13",
@@ -97,7 +110,6 @@
     "eslint-plugin-jsdoc": "~48.2.1",
     "eslint-plugin-prefer-arrow": "~1.2.3",
     "eslint-plugin-unicorn": "^51.0.1",
-    "globby": "^11.1.0",
     "isomorphic-fetch": "~3.0.0",
     "jest": "~29.7.0",
     "jest-junit": "~16.0.0",
