+++ packages/@ama-sdk/schematics/schematics/helpers/open-api-tools-configuration.ts
