+++ packages/@ama-sdk/schematics/schematics/typescript/shell/index.spec.ts
@@ -1,4 +1,4 @@
-import type { OpenApiToolsConfiguration } from '@ama-sdk/core';
+import type { OpenApiToolsConfiguration } from '@ama-sdk/schematics';
 import {Tree} from '@angular-devkit/schematics';
 import { SchematicTestRunner, UnitTestTree } from '@angular-devkit/schematics/testing';
 import * as path from 'node:path';
@@ -99,6 +99,7 @@ describe('Typescript Shell Generator', () => {
   });
 
   it('should generate correct openapitools.json', () => {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-redundant-type-constituents
     const openApiTools = yarnTree.readJson('/openapitools.json') as JsonObject & OpenApiToolsConfiguration;
     // eslint-disable-next-line @typescript-eslint/naming-convention
     expect(openApiTools['generator-cli'].generators).toEqual(expect.objectContaining({'test-scope-test-sdk': expect.anything()}));
