+++ packages/@ama-sdk/schematics/schematics/typescript/shell/index.ts
@@ -52,7 +52,7 @@ function ngGenerateTypescriptSDKFn(options: NgGenerateTypescriptSDKShellSchemati
       'eslint-plugin-prefer-arrow': amaSdkSchematicsPackageJson.devDependencies!['eslint-plugin-prefer-arrow'],
       'eslint-plugin-unicorn': amaSdkSchematicsPackageJson.devDependencies!['eslint-plugin-unicorn'],
       'eslint': amaSdkSchematicsPackageJson.devDependencies!.eslint,
-      'globby': amaSdkSchematicsPackageJson.devDependencies!.globby,
+      'globby': amaSdkSchematicsPackageJson.dependencies!.globby,
       'isomorphic-fetch': amaSdkSchematicsPackageJson.devDependencies!['isomorphic-fetch'],
       'jest': amaSdkSchematicsPackageJson.devDependencies!.jest,
       'rxjs': amaSdkSchematicsPackageJson.dependencies!.rxjs,
