+++ packages/@ama-sdk/schematics/src/public_api.ts
@@ -1,3 +1,7 @@
+export * from '../schematics/helpers/generators';
+export type * from '../schematics/helpers/open-api-tools-configuration';
 export type * from '../schematics/typescript/core/schema';
 export type * from '../schematics/typescript/mock/schema';
 export type * from '../schematics/typescript/shell/schema';
+
+
