+++ packages/@ama-sdk/schematics/tsconfig.cli.json
@@ -10,7 +10,7 @@
   },
   "references": [
     {
-      "path": "./tsconfig.build.json"
+      "path": "./tsconfig.builders.json"
     }
   ],
   "include": [
