+++ packages/@ama-sdk/schematics/tsconfig.json
@@ -7,6 +7,9 @@
     },
     {
       "path": "./tsconfig.spec.json"
+    },
+    {
+      "path": "./tsconfig.cli.json"
     }
   ]
 }
